Allow float values for adding and updating credits

+++ app/api/companies/[id]/credits/update/route.ts
@@ -5,7 +5,7 @@ import { requireAdminSession } from "@/lib/auth/server-session";
 import { updateCredits } from "@/src/server/repositories/setup.repository";
 
 const schema = z.object({
-  delta: z.number().int(),
+  delta: z.number(),
   description: z.string().min(1),
 });
